backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
alembic==1.13.2
psycopg[binary]==3.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
httpx==0.27.2
celery==5.4.0
redis==5.0.8
beautifulsoup4==4.12.3
lxml==5.3.0
